+++ R/cohens_d.R
@@ -1,11 +1,16 @@
 #' Effect size for differences
 #'
-#' Compute different indices of effect size. For very small sample sizes (n < 20) Hedges' g is considered as less biased than Cohen's d. For sample sizes > 20, the results for both statistics are roughly equivalent. The Glass’s delta is appropriate if standard deviations are significantly different between groups, as it uses only the control group's (\code{x}) standard deviation.
+#' Compute different indices of effect size. For very small sample sizes (n < 20)
+#' Hedges' g is considered as less biased than Cohen's d.
+#' For sample sizes > 20, the results for both statistics are roughly equivalent.
+#' \cr\cr
+#'  The Glass’s delta is appropriate if standard deviations are significantly different
+#'  between groups, as it uses only the \emph{second} group's standard deviation.
 #'
-#' @param x A continuous variable or a formula.
-#' @param y A continuous variable, a factor with two groups or a formula.
+#' @param x A formula, a numeric vector, or a name of one in \code{data}.
+#' @param y A numeric vector, a grouping (character / factor) vector, a or a name of one in \code{data}. Ignored if \code{x} is a formula.
 #' @param data An optional data frame containing the variables.
-#' @param correction If \code{TRUE}, applies a correction to the formula to make it less biased for small samples (McGrath & Meyer, 2006).
+#' @param correction If \code{TRUE}, applies a correction to make it less biased for small samples (McGrath & Meyer, 2006).
 #' @param pooled_sd If \code{FALSE}, the regular SD from both combined groups is used instead of the \code{\link{sd_pooled}}.
 #' @param paired If \code{TRUE}, the values of \code{x} and \code{y} are considered as paired.
 #' @inheritParams chisq_to_phi
@@ -22,10 +27,12 @@
 #' @examples
 #' cohens_d(iris$Sepal.Length, iris$Sepal.Width)
 #' hedges_g("Sepal.Length", "Sepal.Width", data = iris)
-#' glass_delta(Sepal.Length ~ Sepal.Width, data = iris)
 #'
-#' cohens_d(iris$Sepal.Length, iris$Sepal.Width, correct = TRUE, pooled_sd = FALSE)
-#' cohens_d(Sepal.Length ~ Species, data = iris[iris$Species %in% c("versicolor", "setosa"), ])
+#' cohens_d(mpg ~ am, data = mtcars)
+#' cohens_d(mpg ~ am, data = mtcars, pooled_sd = FALSE)
+#' hedges_g(mpg ~ am, data = mtcars)
+#' glass_delta(mpg ~ am, data = mtcars)
+#'
 #' @references \itemize{
 #'  \item Cohen, J. (2013). Statistical power analysis for the behavioral sciences. Routledge.
 #'  \item McGrath, R. E., & Meyer, G. J. (2006). When effect sizes disagree: the case of r and d. Psychological methods, 11(4), 386.
@@ -33,7 +40,13 @@
 #' }
 #' @importFrom stats var model.frame
 #' @export
-cohens_d <- function(x, y = NULL, data = NULL, correction = FALSE, pooled_sd = TRUE, paired = FALSE, ci = 0.95) {
+cohens_d <- function(x,
+                     y = NULL,
+                     data = NULL,
+                     correction = FALSE,
+                     pooled_sd = TRUE,
+                     paired = FALSE,
+                     ci = 0.95) {
   .effect_size_difference(
     x,
     y = y,
@@ -48,7 +61,13 @@ cohens_d <- function(x, y = NULL, data = NULL, correction = FALSE, pooled_sd = T
 
 #' @rdname cohens_d
 #' @export
-hedges_g <- function(x, y = NULL, data = NULL, correction = FALSE, pooled_sd = TRUE, paired = FALSE, ci = 0.95) {
+hedges_g <- function(x,
+                     y = NULL,
+                     data = NULL,
+                     correction = FALSE,
+                     pooled_sd = TRUE,
+                     paired = FALSE,
+                     ci = 0.95) {
   .effect_size_difference(
     x,
     y = y,
@@ -76,51 +95,64 @@ glass_delta <- function(x, y = NULL, data = NULL, correction = FALSE, ci = 0.95)
 
 
 
-
-
-
-
-
-
-
 #' @importFrom stats sd
 #' @keywords internal
-.effect_size_difference <- function(x, y = NULL, data = NULL, type = "d", correction = FALSE, pooled_sd = TRUE, paired = FALSE, ci = ci) {
+.effect_size_difference <- function(x,
+                                    y = NULL,
+                                    data = NULL,
+                                    type = "d",
+                                    correction = FALSE,
+                                    pooled_sd = TRUE,
+                                    paired = FALSE,
+                                    ci = 0.95) {
   out <- .deal_with_cohens_d_arguments(x, y, data)
   x <- out$x
   y <- out$y
 
-  if (paired) {
-    n <- length(x)
-    df <- n - 1
-  } else {
-    n <- length(c(x,y))
-    df <- n - 2
+  if (is.null(y)) {
+    if (type == "delta") {
+      stop("For Glass' Delta, please provide data from two samples.", call. = FALSE)
+    }
+    y <- 0
+    paired <- TRUE
   }
 
   # Compute index
-  diff_of_means <- mean(x, na.rm = TRUE) - mean(y, na.rm = TRUE)
-
-  if (type == "d" | type == "g") {
-    if (paired) {
-      denominator <- stats::sd(x - y, na.rm = TRUE)
-    } else {
+  if (paired) {
+    d <- mean(x - y, na.rm = TRUE)
+    s <- stats::sd(x - y, na.rm = TRUE)
+    n <- length(x)
+    df <- n - 1
+    t <- d / (s / sqrt(n))
+  } else {
+    d <- mean(x, na.rm = TRUE) - mean(y, na.rm = TRUE)
+    n1 <- length(x)
+    n2 <- length(y)
+    n <- n1 + n2
+    df <- n - 2
+    if (type == "d" | type == "g") {
       if (pooled_sd) {
-        denominator <- sd_pooled(x, y)
+        s <- suppressWarnings(sd_pooled(x, y))
+        t <- d / (s * sqrt(1 / n1 + 1 / n2))
       } else {
-        denominator <- stats::sd(c(x, y), na.rm = TRUE)
+        s1 <- stats::sd(x, na.rm = TRUE)
+        s2 <- stats::sd(y, na.rm = TRUE)
+        s <- sqrt((s1 ^ 2 + s2 ^ 2) / 2)
+        # s <- stats::sd(c(x, y), na.rm = TRUE)
+        t <- d / sqrt(s1 ^ 2 / n1 + s2 ^ 2 / n2)
       }
+    } else if (type == "delta") {
+      s <- stats::sd(y, na.rm = TRUE)
+      t <- d / (s * sqrt(1 / n1 + 1 / n2))
     }
-  } else if (type == "delta") {
-    denominator <- stats::sd(x, na.rm = TRUE)
   }
 
-  out <- data.frame(d = diff_of_means / denominator)
+  out <- data.frame(d = d / s)
   types <- c("d" = "Cohens_d", "g" = "Hedges_g", "delta" = "Glass_delta")
   colnames(out) <- types[type]
   out <- cbind(
     out,
-    .ci_from_t_test(x, y, type = type, paired = paired, ci = ci)
+    t_to_d(t, df, ci = ci, paired = paired)[,-1 , drop = FALSE]
   )
 
   if (type == "g") {
@@ -135,10 +167,11 @@ glass_delta <- function(x, y = NULL, data = NULL, correction = FALSE, ci = 0.95)
   }
 
   # McGrath & Meyer (2006)
-  if (correction == TRUE) {
+  if (correction) {
+    correction <- ((n - 3) / (n - 2.25)) * sqrt((n - 2) / n)
+
     out[, colnames(out) %in% c(types, "CI_low", "CI_high")] <-
-      out[, colnames(out) %in% c(types, "CI_low", "CI_high")] *
-      ((n - 3) / (n - 2.25)) * sqrt((n - 2) / n)
+      out[, colnames(out) %in% c(types, "CI_low", "CI_high")] * correction
   }
 
   class(out) <- c("effectsize_table", class(out))
@@ -147,6 +180,10 @@ glass_delta <- function(x, y = NULL, data = NULL, correction = FALSE, ci = 0.95)
 
 
 
+
+
+
+# Utils -------------------------------------------------------------------
 
 
 
@@ -160,13 +197,23 @@ glass_delta <- function(x, y = NULL, data = NULL, correction = FALSE, ci = 0.95)
     }
   }
 
+  ## Preprocess data
 
-  # Preprocess data
+  # Formula
   if (inherits(x, "formula")) {
-    data <- model.frame(x, data = data)
-    x <- names(data)[1]
-    y <- names(data)[2]
+    trms <- terms(x)
+
+    group <- all.vars(delete.response(trms))
+    outcome <- setdiff(all.vars(trms),group)
+
+    if (!(length(outcome) == 1 & length(group) == 1)) {
+      stop("Formula must have the 'outcome ~ group'.", call. = FALSE)
+    }
+
+    x <- data[[outcome]]
+    y <- as.factor(data[[group]])
   }
+
 
   if (is.character(x)) {
     x <- data[[x]]
@@ -176,41 +223,31 @@ glass_delta <- function(x, y = NULL, data = NULL, correction = FALSE, ci = 0.95)
     y <- data[[y]]
   }
 
+  if (!is.numeric(x)) {
+    stop("Cannot compute effect size for a non-numeric vector.", call. = FALSE)
+  }
 
   # If y is a factor
   if (!is.null(y)) {
-    if (is.factor(y) | is.character(y)) {
-      if (length(x) != length(y)) {
-        stop("The length of the group factor must be the same.")
-      }
+    if (!is.numeric(y)) {
       if (length(unique(y)) > 2) {
-        stop("Cannot compute the difference as a factor with more than 2 levels has been provided.")
-      } else {
-        groups <- as.character(y)
-        y <- x[groups == unique(groups)[2]]
-        x <- x[groups == unique(groups)[1]]
+        stop("Cannot compute the difference as a factor with more than 2 levels has been provided.",
+             call. = FALSE)
       }
+      if (length(x) != length(y)) {
+        stop("Grouping variable must be the same length.", call. = FALSE)
+      }
+
+      data <- split(x, y)
+      x <- data[[1]]
+      y <- data[[2]]
+    } else if (length(unique(y)) == 2) {
+      warning(
+        "'y' is numeric but has only 2 unique values. If this is a grouping variable, convert it to a factor.",
+        call. = FALSE
+      )
     }
   }
 
   list(x = x, y = y)
-}
-
-#' @keywords internal
-.ci_from_t_test <- function(x, y, type = "d", paired = FALSE, ci = 0.95){
-  if (paired) {
-    tobj <- stats::t.test(x - y)
-    df <- length(x) - 1
-  } else {
-    if (type == "delta") {
-      tobj <- stats::t.test(x, mu = mean(y))
-      df <- length(x) - 1
-      paired <- TRUE
-    } else {
-      tobj <- stats::t.test(x, y, var.equal = TRUE)
-      df <- length(c(x,y)) - 2
-    }
-  }
-
-  t_to_d(unname(tobj$statistic), df, ci = ci, pooled = paired)[, -1 , drop = FALSE]
 }