+++ R/convert_tFz_to_d.R
@@ -1,10 +1,18 @@
 #' @rdname t_to_r
 #' @export
-t_to_d <- function(t, df_error, pooled = FALSE, ci = 0.95, ...) {
-  # Will be 1 if TRUE, and 2 if FALSE
-  pooled <- 2 - pooled
+t_to_d <- function(t, df_error, paired = FALSE, ci = 0.95, pooled, ...) {
+  if (!missing(pooled)) {
+    paired <- pooled
+    warning(
+      "Argument 'pooled' is deprecated, use 'paired' instead. Setting paired <- pooled.",
+      call. = FALSE
+    )
+  }
 
-  res <- data.frame(d = pooled * t / sqrt(df_error))
+  # Will be 1 if TRUE, and 2 if FALSE
+  paired <- 2 - paired
+
+  res <- data.frame(d = paired * t / sqrt(df_error))
 
   if (is.numeric(ci)) {
     stopifnot(length(ci) == 1, ci < 1, ci > 0)
@@ -13,8 +21,8 @@ t_to_d <- function(t, df_error, pooled = FALSE, ci = 0.95, ...) {
     ts <- t(mapply(.get_ncp_t,
                    t, df_error, ci))
 
-    res$CI_low <- pooled * ts[,1] / sqrt(df_error)
-    res$CI_high <- pooled * ts[,2] / sqrt(df_error)
+    res$CI_low <- paired * ts[,1] / sqrt(df_error)
+    res$CI_high <- paired * ts[,2] / sqrt(df_error)
   }
 
   class(res) <- c("effectsize_table", class(res))
@@ -34,11 +42,19 @@ convert_t_to_d <- t_to_d
 
 #' @rdname t_to_r
 #' @export
-z_to_d <- function(z, n, pooled = FALSE, ci = 0.95, ...) {
-  # Will be 1 if TRUE, and 2 if FALSE
-  pooled <- 2 - pooled
+z_to_d <- function(z, n, paired = FALSE, ci = 0.95, pooled, ...) {
+  if (!missing(pooled)) {
+    paired <- pooled
+    warning(
+      "Argument 'pooled' is deprecated, use 'paired' instead. Setting paired <- pooled.",
+      call. = FALSE
+    )
+  }
 
-  res <- data.frame(d = pooled * z / sqrt(n))
+  # Will be 1 if TRUE, and 2 if FALSE
+  paired <- 2 - paired
+
+  res <- data.frame(d = paired * z / sqrt(n))
 
   if (is.numeric(ci)) {
     stopifnot(length(ci) == 1, ci < 1, ci > 0)
@@ -50,8 +66,8 @@ z_to_d <- function(z, n, pooled = FALSE, ci = 0.95, ...) {
     qs <- qnorm(probs)
     zs <- cbind(qs[1] + z, qs[2] + z)
 
-    res$CI_low <- pooled * zs[,1] / sqrt(n)
-    res$CI_high <- pooled * zs[,2] / sqrt(n)
+    res$CI_low <- paired * zs[,1] / sqrt(n)
+    res$CI_high <- paired * zs[,2] / sqrt(n)
   }
 
   class(res) <- c("effectsize_table", class(res))
@@ -71,11 +87,19 @@ convert_z_to_d <- z_to_d
 
 #' @rdname t_to_r
 #' @export
-F_to_d <- function(f, df, df_error, pooled = FALSE, ci = 0.95, ...) {
+F_to_d <- function(f, df, df_error, paired = FALSE, ci = 0.95, pooled, ...) {
+  if (!missing(pooled)) {
+    paired <- pooled
+    warning(
+      "Argument 'pooled' is deprecated, use 'paired' instead. Setting paired <- pooled.",
+      call. = FALSE
+    )
+  }
+
   if (df > 1) {
     stop("Cannot convert F with more than 1 df to (partial) r.")
   }
-  t_to_d(sqrt(f), df_error, pooled, ci)
+  t_to_d(sqrt(f), df_error, paired, ci)
 }
 
 #' @rdname t_to_r