+++ tests/testthat/test-standardize_parameters.R
@@ -1,12 +1,10 @@
-context("standardize_parameters")
+if (require("testthat") && require("effectsize")) {
+  test_that("standardize_parameters (simple)", {
+    df <- iris
+    r <- as.numeric(cor.test(df$Sepal.Length, df$Petal.Length)$estimate)
 
-
-
-test_that("standardize_parameters (simple)", {
-  # df <- iris
-  # r <- as.numeric(cor.test(df$Sepal.Length, df$Petal.Length)$estimate)
-  #
-  # model <- lm(Sepal.Length ~ Petal.Length, data = df)
-  # es <- standardize_parameters(model)[2, 2]
-  # testthat::expect_equal(es, r, tol = 0.01)
-})
+    model <- lm(Sepal.Length ~ Petal.Length, data = df)
+    es <- standardize_parameters(model)[2, 2]
+    testthat::expect_equal(es, r, tol = 0.01)
+  })
+}